Differentiate expressions

Differentiate expressions in the mcstate dsl. This function is exported for advanced use, and really so that we can use it from odin. But it has the potential to be generally useful, so while we'll tweak the interface quite a lot over the next while it is fine to use if you can handle some disruption.

Usage

mcstate_differentiation()

Value

A list of related objects:

differentiate: A function that can differentiate an expression with respect to a variable (as a string).

maths: Some mathmatical utilities for constructing expressions. This will be documented later, but the most useful bits on here are the function elements times, plus and plus_fold.

We will expand this soon to advertise what functions we are able to differentiate to allow programs to fail fast.

Details

R already has support for differentiating expressions using D, which is useful for creating derivatives of simple functions to pass into nonlinear optimisation. We need something a bit more flexible for differentiating models in the mcstate dsl (mcstate_dsl) and also in the related odin dsl.

Differences to D()

We try a little harder to simplify expressions.

The distribution functions in the mcstate DSL (e.g., Poisson) are (will be) handled specially, allowing substitution of log-densities and expectations.

Once we support array expressions, we will be able to differentiate through these.

Roadmap

We may need to make this slightly extensible in future, but for now the set of functions that can be differentiated is closed.