Information about supported distributions

Report information about supported distributions in the dsl. This is primarily intended for use in packages which use mcstate_dsl_parse_distribution, as this function reports information about which distributions and arguments would succeed there.

Usage

mcstate_dsl_distributions()

Value

A data.frame with columns

name the name of the distribution; each name begins with a capital letter, and there are duplicate names where different parameterisations are supported.

args the arguments of all parameters, except the random variable itself which is given as the first argument to density functions.

We may expand the output here in the future to include information on if distributions have support in C++, but we might end up supporting everything this way soon.