Compute gradient of log density

Compute the gradient of log density (which is returned by mcstate_model_density) with respect to parameters. Not all models support this, and an error will be thrown if it is not possible.

Usage

mcstate_model_gradient(model, parameters, named = FALSE)

Arguments

model

An mcstate_model object

parameters

A vector or matrix of parameters

named

Logical, indicating if the output should be named using the parameter names.

Value

A vector or matrix of gradients

See also