Run MCMC chain in series

Run MCMC chains in series (one after another). This is the simplest chain runner, and the default used by mcstate_sample(). It has nothing that can be configured (yet).

Usage

mcstate_runner_serial(progress = NULL)

Arguments

progress

Optional logical, indicating if we should print a progress bar while running. If NULL, we use the value of the option mcstate2.progress if set, otherwise we show the progress bar (as it is typically wanted). The progress bar itself responds to cli's options; in particular cli.progress_show_after and cli.progress_clear will affect your experience.

Value

A runner of class mcstate_runner that can be passed to mcstate_sample()