Domain Specific Language for mcstate

Create a model using the mcstate2 DSL; this function will likely change name in future, as will its interface.

Usage

mcstate_dsl(x, type = NULL)

Arguments

x

The model as an expression. This may be given as an expression, as a string, or as a path to a filename. Typically, we'll do a reasonable job of working out what you've provided but use the type argument to disambiguate or force a particular interpretation. The argument uses rlang's quosures to allow you to work with expressions directly; see examples for details.

type

Force interpretation of the type of expression given as x. If given, valid options are expression, text or file.

Value

A mcstate_model object derived from the expressions you provide.

Examples

# Expressions that correspond to models can be passed in with no
# quoting
mcstate_dsl(a ~ Normal(0, 1))
#> $model
#> $model$parameters
#> [1] "a"
#> 
#> $model$density
#> function (x) 
#> {
#>     pars <- packer$unpack(x)
#>     density <- numeric()
#>     density[["a"]] <- (function (x, mean, sd) 
#>     dnorm(x, mean, sd, log = TRUE))(pars[["a"]], 0, 1)
#>     sum(density)
#> }
#> <environment: 0x56502e4c9b70>
#> 
#> $model$direct_sample
#> function (rng) 
#> {
#>     pars <- list()
#>     pars[["a"]] <- (function (rng, mean, sd) 
#>     rng$normal(1, mean, sd))(rng, 0, 1)
#>     unlist(pars[packer$parameters], FALSE, FALSE)
#> }
#> <environment: 0x56502e4c9b70>
#> 
#> 
#> $parameters
#> [1] "a"
#> 
#> $parameter_groups
#> NULL
#> 
#> $domain
#>   [,1] [,2]
#> a -Inf  Inf
#> 
#> $density
#> function (x) 
#> {
#>     pars <- packer$unpack(x)
#>     density <- numeric()
#>     density[["a"]] <- (function (x, mean, sd) 
#>     dnorm(x, mean, sd, log = TRUE))(pars[["a"]], 0, 1)
#>     sum(density)
#> }
#> <environment: 0x56502e4c9b70>
#> 
#> $gradient
#> NULL
#> 
#> $direct_sample
#> function (rng) 
#> {
#>     pars <- list()
#>     pars[["a"]] <- (function (rng, mean, sd) 
#>     rng$normal(1, mean, sd))(rng, 0, 1)
#>     unlist(pars[packer$parameters], FALSE, FALSE)
#> }
#> <environment: 0x56502e4c9b70>
#> 
#> $rng_state
#> NULL
#> 
#> $properties
#> $has_gradient
#> [1] FALSE
#> 
#> $has_direct_sample
#> [1] TRUE
#> 
#> $is_stochastic
#> [1] FALSE
#> 
#> $has_parameter_groups
#> [1] FALSE
#> 
#> $allow_multiple_parameters
#> [1] FALSE
#> 
#> attr(,"class")
#> [1] "mcstate_model_properties"
#> 
#> attr(,"class")
#> [1] "mcstate_model"
mcstate_dsl({
  a ~ Normal(0, 1)
  b ~ Exponential(1)
})
#> $model
#> $model$parameters
#> [1] "a" "b"
#> 
#> $model$density
#> function (x) 
#> {
#>     pars <- packer$unpack(x)
#>     density <- numeric()
#>     density[["a"]] <- (function (x, mean, sd) 
#>     dnorm(x, mean, sd, log = TRUE))(pars[["a"]], 0, 1)
#>     density[["b"]] <- (function (x, rate) 
#>     dexp(x, rate, log = TRUE))(pars[["b"]], 1)
#>     sum(density)
#> }
#> <environment: 0x56502c38cff8>
#> 
#> $model$direct_sample
#> function (rng) 
#> {
#>     pars <- list()
#>     pars[["a"]] <- (function (rng, mean, sd) 
#>     rng$normal(1, mean, sd))(rng, 0, 1)
#>     pars[["b"]] <- (function (rng, rate) 
#>     rng$exponential(1, rate))(rng, 1)
#>     unlist(pars[packer$parameters], FALSE, FALSE)
#> }
#> <environment: 0x56502c38cff8>
#> 
#> 
#> $parameters
#> [1] "a" "b"
#> 
#> $parameter_groups
#> NULL
#> 
#> $domain
#>   [,1] [,2]
#> a -Inf  Inf
#> b -Inf  Inf
#> 
#> $density
#> function (x) 
#> {
#>     pars <- packer$unpack(x)
#>     density <- numeric()
#>     density[["a"]] <- (function (x, mean, sd) 
#>     dnorm(x, mean, sd, log = TRUE))(pars[["a"]], 0, 1)
#>     density[["b"]] <- (function (x, rate) 
#>     dexp(x, rate, log = TRUE))(pars[["b"]], 1)
#>     sum(density)
#> }
#> <environment: 0x56502c38cff8>
#> 
#> $gradient
#> NULL
#> 
#> $direct_sample
#> function (rng) 
#> {
#>     pars <- list()
#>     pars[["a"]] <- (function (rng, mean, sd) 
#>     rng$normal(1, mean, sd))(rng, 0, 1)
#>     pars[["b"]] <- (function (rng, rate) 
#>     rng$exponential(1, rate))(rng, 1)
#>     unlist(pars[packer$parameters], FALSE, FALSE)
#> }
#> <environment: 0x56502c38cff8>
#> 
#> $rng_state
#> NULL
#> 
#> $properties
#> $has_gradient
#> [1] FALSE
#> 
#> $has_direct_sample
#> [1] TRUE
#> 
#> $is_stochastic
#> [1] FALSE
#> 
#> $has_parameter_groups
#> [1] FALSE
#> 
#> $allow_multiple_parameters
#> [1] FALSE
#> 
#> attr(,"class")
#> [1] "mcstate_model_properties"
#> 
#> attr(,"class")
#> [1] "mcstate_model"

# You can also pass strings
mcstate_dsl("a ~ Normal(0, 1)")
#> $model
#> $model$parameters
#> [1] "a"
#> 
#> $model$density
#> function (x) 
#> {
#>     pars <- packer$unpack(x)
#>     density <- numeric()
#>     density[["a"]] <- (function (x, mean, sd) 
#>     dnorm(x, mean, sd, log = TRUE))(pars[["a"]], 0, 1)
#>     sum(density)
#> }
#> <environment: 0x56502c1e5350>
#> 
#> $model$direct_sample
#> function (rng) 
#> {
#>     pars <- list()
#>     pars[["a"]] <- (function (rng, mean, sd) 
#>     rng$normal(1, mean, sd))(rng, 0, 1)
#>     unlist(pars[packer$parameters], FALSE, FALSE)
#> }
#> <environment: 0x56502c1e5350>
#> 
#> 
#> $parameters
#> [1] "a"
#> 
#> $parameter_groups
#> NULL
#> 
#> $domain
#>   [,1] [,2]
#> a -Inf  Inf
#> 
#> $density
#> function (x) 
#> {
#>     pars <- packer$unpack(x)
#>     density <- numeric()
#>     density[["a"]] <- (function (x, mean, sd) 
#>     dnorm(x, mean, sd, log = TRUE))(pars[["a"]], 0, 1)
#>     sum(density)
#> }
#> <environment: 0x56502c1e5350>
#> 
#> $gradient
#> NULL
#> 
#> $direct_sample
#> function (rng) 
#> {
#>     pars <- list()
#>     pars[["a"]] <- (function (rng, mean, sd) 
#>     rng$normal(1, mean, sd))(rng, 0, 1)
#>     unlist(pars[packer$parameters], FALSE, FALSE)
#> }
#> <environment: 0x56502c1e5350>
#> 
#> $rng_state
#> NULL
#> 
#> $properties
#> $has_gradient
#> [1] FALSE
#> 
#> $has_direct_sample
#> [1] TRUE
#> 
#> $is_stochastic
#> [1] FALSE
#> 
#> $has_parameter_groups
#> [1] FALSE
#> 
#> $allow_multiple_parameters
#> [1] FALSE
#> 
#> attr(,"class")
#> [1] "mcstate_model_properties"
#> 
#> attr(,"class")
#> [1] "mcstate_model"